fix(#2131): split over-cap test files for FR-006 ratchet gate
ci-simulate.test.ts (380→409, was 553) and sdd.test.ts (1078→1078, was 1150)
exceeded the 500-line ratchet. Extract junction/realpathSync tests into
ci-simulate-junction.test.ts and allowedMutationsFor tests into
sdd-mutations.test.ts.
